1. 폴더구조 chat으로 전부 변경
