+++ src/main/java/com/jinan/profile/repository/ChatRoomRepository.java
@@ -1,6 +1,6 @@
 package com.jinan.profile.repository;
 
-import com.jinan.profile.domain.message.ChatRoom;
+import com.jinan.profile.domain.chat.ChatRoom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ChatRoomRepository extends JpaRepository<ChatRoom, Long> {
